cookbooks.sre.dns: add roll-reboot cookbook
Add a cookbook for the rolling reboots of DNS hosts. This will allow us
to perform reboots of the DNS hosts automatically instead of the current
manual process and was made possible by the changes in T347054, where we
moved the state management of DNS hosts to confd.

There is no restart_daemons and that is intentional: we don't need to
restart gdnsd, and for pdns-recursor restarts, we handle them manually.

Bug: T347054
Change-Id: Ib85ea7d9e0c4e72839110fc58c882e4676b278a9