roll-restart-reboot-durum: don't disable Puppet on restart
In the current state of the cookbook, the bird service is not restarted
automatically because Puppet doesn't restart the daemons. Since there is
a systemd ordering in place, we should not be worried about that and
instead let nginx.service restart what is required. This is similar to
what we observed in I0b5107a0, so copy that fix over to the durum
cookbook as well.

Change-Id: I52e46b7ee717a514a21b751be68eedee9c2f8415