wdqs.data-reload: support HDFS as a source
Add the ability to source the dump from HDFS by using a stat node as an
intermediary:
- hdfs-rsync from HDFS to a local path on a stat machine
- transferfy from that stat machine to the wdqs node
- continue with the usual loadData process, skipping the munge operation

Added 3 new profiles:
- wikidata-full: the full graph but loaded from HDFS
- wikidata-main
- wikidata-scholarly

Refactor the cookbook to have the same logic between commons & wikidata.
Fix the handling of the kafka timestamp that was passed as a string
instead of a int.

Switch to CookbookBase so that the various operation are better isolated
in an encapsulating class.

Drop the ability to specify the path of the lexeme&full dumps it was
apperently only used by the graph_split profile.

Bug: T349069
Change-Id: I09a8e0cb8609f8a130fc87163b9b49b911c953b3