sre.host.provision: no-op refactor to highlight DELL-specific confs
This is a proposal to start splitting the common bits from
the vendor-specific ones. This would be the first step
to pave the way to add the initial support for Supermicro,
that will likely be incremental over time.
The idea is to:
- Keep the current workflow for Dell nodes, that is battle
  tested and that works using the SCP endpoint.
- Add the support for Supermicro, that will likely be different,
  and work with DCops to get to a good state to help them
  provisioning all the upcoming Supermicro hosts that have
  been ordered.
Once we have a stable cookbook supporting both (or even while
adding Supermicro bits) we have the option to refactor/unify
common procedures and code. Starting to do a refactor of Dell's
specific code before even having anything that works with
Supermicro may end up in premature optimization, and a ton of
time invested in changing something that already works (imho).

Bug: T365372
Change-Id: I3ce81d386ee300adcbd0b5e9696e5f939533abd3