sre.hadoop.reboot-workers: use ceil not floor
ISSUE

erroneous use of math.floor instead of math.ceil that
results in batch size not being respected
in some circumstances

CONTEXT

This cookbook appears to want the following behavior:

Given a list of hosts, split it into x batches of size n,
with each batch having no more than n hosts.

However the way it was written, with floor instead of ceiling,
meant that you could end up with a batch with > n hosts.

For example, with 106 total hosts and batch size 3, the logic
could return 4 hosts in the first batch instead of 3.

Bug: T367592
Change-Id: I4d1cc17a97daf7c34c7ebdbcb84b7b49392dc46e