cookbooks/sre/dns: add a cookbook for roll restart of ntpd.service
Puppet no longer manages the restarts of the ntpd.service for us -- an
intentional change since we want to control the time between the hosts
and allow for proper NTP sync. So far we have been using a cumin command
to do the restarts of ntpd.service in case the ntp.conf file is changed.

This cookbook helps us do that automatically without worrying or knowing
about which command to run, the sane batch defaults, etc.

Change-Id: If64ce078ddce1953d6f98b801600c07605e83a08