sre.gitlab.upgrade: lock backups during upgrades
GitLab upgrades can be disrupted by backups running
concurrent to the upgrade. The apt package upgrade is doing a database
backup automatically which fails when another backup is running.
So this change keeps the lock after a successful backup and removes
it at the end of the cookbook. I moved the methodes to __init__ because
they are used by the upgrade and failover cookbooks.

Bug: T367501
Change-Id: I6db0f073e40ac0929a25afb3ad8a22e68dc6a8f7