sre.cdn.roll-reboot: log host to SAL post reboot
The Traffic team uses the sre.cdn.roll-reboot to reboot the CDN and we
usually do it on the entire site for both text and upload. Given that
there is a long interval in between each host, we have no idea which
host was rebooted or where the cookbook progress is currently unless we
check the screen session of the user who started the cookbook.

We operate in batch of 1 for this so whenever a host is rebooted, we
should log it to SAL so we know which host is done rebooting and in
general, to increase the verbosity of the cookbook operation.

Change-Id: I3fd45c2e0a8c30d9043ccb2fd1a7d9e4af1a8730