wdqs: store metadata about graph split type
With the changes coming in the WDQS graph split epic,
we no longer have only 1 possible journal type. As such
the data-reload and data-transfer cookbooks need to store
and read metadata about which subgraph type a given host's
journal file corresponds to.

In a future patch we'll make the data transfer cookbook refuse to
transfer the file if there is a graph type mismatch, unless the
operator supplies an override flag argument.

OTHER CHANGES

- only sleep if we actually performed an lvs action

Bug: T364077
Change-Id: Ib10f22cd1f36e174d7bbea34876f81f1882040a1