sre.network.tls: use a different client certificate to authenticate
After a lot of digging it seems that we need to send the whole
chain of intermediate + client certificate to be able to allow
cfssl gencert to work with the httpd settings. This is the same
configuration that we have in puppet, see profile::pki::client.

More technical background in T340557#8985560

Bug: T355750
Change-Id: I13e269f9a9c452f45fb6459354750b14abeb64cb