switchdc: mediawiki cache warmup now targets k8s
With this change, the cache warmup switchdc cookbook now invokes the
newly updated k8s-aware warmup script on the active deployment host.

This adopts the same idiom as other cookbooks that need to run remote
commands on the active deployment host.

There is ongoing discussion on spicerack offering an accessor for this
purpose, in which case, this cookbook should be updated.

Bug: T369921
Depends-On: Ic48417e5acb0a64cd6af1c66a2b25853a8c2a5ef
Depends-On: Ieb907d6f13bcfb35cab9609685552e1a5184cc39
Change-Id: I5e706b4cd2875117945cdd867aec5a763f1339c5