Remove clear_dhcp_cache function from reimage cookbook
Thankfully we have now upgraded all the EVPN switches in Eqiad
rows E and F to JunOS 22.2.  The newer version gives us the option
to configure the DHCP relay function in a stateless "forward" only
mode.  In this mode the switch does not create a DHCP binding entry,
the expiry of which was previously triggering the bug that blocked
ARP for newly reimaged hosts.

TL;DR we no longer need this workaround to manually clear the DHCP
binding entry and ARP/L2 to prevent the bug from being triggered.

Bug: T306421
Change-Id: I58e3995a782177f93bf0f689e9461bdbfffdf326