sre.hosts.provision: fix dell_config_changes
The dell_config_change property was added as part of a refactoring,
but it was wrong. The cookbook assumes that the dictionary can
change and maintain state, meanwhile using the property this
is not possible. We are currently missing some options like console
redirect due to that (on newly provisioned hw).

Bug: T365372
Change-Id: Ifecf12c6eb2a672265bbae3162eed84021744a50