sre.dns.admin: add cookbook for GeoDNS pool/depool
This is the first commit of the cookbook for DNS depooling of sites/geo
DNS. It is meant to serve as a wrapper for confctl and is intended to be
the primary method of performing administrative actions (pool/depool) on
sites by SREs.

Once the related changes that this cookbook depends on are merged, all
state management for admin_state will happen through this cookbook and
not through Git (the current approach).

Bug: T369366
Depends-On: I310affc866062f3763fb3e378d763332e36a0f33
Change-Id: If45adac9b040cf3f8d30ebc35101005d41ab3ed2