Update confd template error state globs
In T363924, the error state file written by the confd-lint-wrap script
was updated to match the unambiguous name of the underlying template,
rather than mirroring the name of the rendered template tempfile.

This was done in order to address spurious alerts when two resources
have identical basenames (i.e., stored at different paths).

In any case, this means we now need to look for files like:

  _var_lib_gdnsd_discovery-mw-api-int-ro.state.err

rather than:

  .discovery-mw-api-int-ro.stateNNNNNN.err

The fact that these cookbooks were not updated a couple months ago was
an oversight.

Bug: T363924
Bug: T372649
Change-Id: I609abac20d201e57e2eef9afe6340c9b9d77cd20