sre.switchdc.mediawiki: add --task-id argument
The task ID is used in two contexts:
1. It's included in the admin reason, used in, e.g., downtimes.
2. It's now used to drive task comment updates recording cookbook
   completion and success / failure status.

In order to achieve #2, MediaWikiSwitchDCRunnerBase now implements run,
and the action of individual cookbooks in delegated to action, which is
to be implemented by subclass cookbooks.

While conceptually similar to the role of CookbookRunnerBase.rollback,
that's not quite what we want, as we want to capture additional context
info (e.g., elapsed time).

Bug: T330273
Change-Id: Idbd11f15e4f6ec1abbf5bdcc6975973f8155aaa1