kafka/roll-restart-reboot-brokers: Add exclude and no-election options
When replacing kafka brokers it would be nice to be able to use the
restart functionality of this cookbook. This requires the exclude option
as kafka will already be stopped on the to be replaced node at time of
kafka restarts and we don't want to start it again.

It also seemed desireable not to run the preferred-replica-election for
every restart but call it manually after the new broker is in sync.

Bug: T373189
Change-Id: I9e3a19711bc23a0aa4f83543f2d48c3f1429d743