sre.switchdc.mediawiki: show TTL sleep end time
Unless you're watching closely as the 00-reduce-ttl cookbook runs, it
can be unclear when the TTL sleep at the end will complete.

While one could search the cookbook logs to check the timestamp, it's a
lot simpler to just say when we'll be done sleeping in the message.

Now does so, with HH:MM:SS granularity and always in UTC.

Bug: T374047
Change-Id: I103b5c98f64806d72d03c35280a34dbb12312de5