sre.dns.admin: add guardrails for depool of sites/resources
Since this cookbook is the only way for SREs to depool/pool sites going
forward, add some safety checks to ensure that depool actions meet a
certain threshold, both in the number of sites that can be depooled at
once (for all resources) and the locations where the sites can be
depooled. This is from the perspective of the Traffic team.

This is a first attempt at having some sort of safety checks to prevent
SREs from depooling more resources than desirable. We will amend this
based on feedback, if any.

Change-Id: I25e1ae1bd66a01db9a2b1374c05a65b871617aaf