wdqs.data-transfer: refuse xfer on differing jnl
We don't want to allow transferring from a
wdqs-main host to wdqs-scholarly or vice-versa.

This implementation checks that the contents of
the data_loaded file, which specifies the precise graph type,
matches on the source and dest host.

This behavior can be disabled with
--no-check-graph-type
which is intended to be used when intializing new graph hosts.

Bug: T364077
Change-Id: I703552ee62820317cdf43d13a8b65421e9fa64e1