sre.discovery.datacenter: exclude swift-https
During the September 2024 switchover, thumbor in codfw became overloaded
after swift.discovery.wmnet was depooled in eqiad.

We need a better understood capacity model for this service (possibly
together with the swift frontends), and then to validate that model
empirically, before swift is considered safe to "electively" depool as
part of switchover day 1.

Bug: T370962
Bug: T376516
Change-Id: I323ab52703c0cce2c142a8c04c675a1f8561ad50