wdqs.categories-reload: don't check host
This check mainly makes sense in the context of
a data-transfer where we're transferring from
a source to a dest and want to sanity check
that the host is part of a WDQS fleet.

In the categories reload world we're only
running this on 1 host.

Furthermore we're (likely) moving categories
away from WDQS hosts to elsewhere (either VM or k8s)
so some of this logic will be going away regardless.

Also, this check is preventing us from
doing an experimental categories reload on a VM.

Bug: T375687
Change-Id: I9f0a81c7bd2b12c4df1b83ac5170568e8e61948e