sre.switchdc.mediawiki: add mwdebug-next to MEDIAWIKI_SERVICES
While the motivation for creating this service is the PHP 8.1 migration
(which should be done well before the next switchover), we plan to keep
it around for use in supporting future migrations.

Note: Technically mwdebug* are configured as active-active services, but
we currently treat them as if they were active-passive.

Bug: T372604
Change-Id: I797091c1a86ae5c22393b8918c1af85d173311f4