sre.hosts.provision: first refactor with vendor-specific classes
The provision cookbook became difficult to read and manage after
the introduction of the Supermicro bits, since most of the code
is now full of if-branches related to what vendor is the server to
be configured from.

This is a first attempt to refactor the code, and the idea is simple:
use spicerack's data in the Provision class to create vendor-specific
runner classes.

Some code is still duplicated, imho further attempts to DRY the code should
be done (if we think so) afterwards to simplify the review of
the various refactoring attempts (this is already complicated).

Bug: T365372
Change-Id: I742cf860ff620b566118231567085eb8f042de44