sre.hosts.provision: raise RuntimeError if Redfish returns an error
Explicitly fail with an exception, when applying settings to a Supermicro's BMC,
if Redfish returns a non 2XX response.

Typical use case that we are facing: the wrong password/config/etc..
is pushed via Redfish and an error like Unauthorized (400) is
returned. The cookbook displays the error and keeps going, so it is
easy for the operator to miss the fact that something went wrong.
For example, if the network config of the BMC is not applied cleanly
the connectivity will be present until the DHCP lease expires, making
it like something is working while it is only temporary.

Bug: T365372
Change-Id: I94281509b81f1fb3b37c5bd4e043a4d9c823200d