sre.hosts.provision: fix hw raid detection for Supermicro
The ms-be20XX Supermicro nodes that we are setting up have
a 'HA-RAID' controller deployed and available via Redfish,
not a simple 'RAID' one, so the startswith() check fails.
I am not sure if Supermicro offers a 'HA-RAID' and 'RAID',
in case a different check would be needed, but it is probably
better to allow a broader scope only if really needed
in the future.

Bug: T371400
Change-Id: I1ceb748866058b240ceac5dcf5fa634f65d165cd