sre.hosts.reimage: clear puppetdb's state upon rollback (if needed)
While reimaging ms-be2XXXX nodes we ended up in this situation:
T371400#10236404

My understanding of the issue is the following:
- ms-be hosts automatically get the production role/profiles,
  they don't go through insetup::datapersistence. This is convenient
  for DP since most of the checks for disks are added only when
  the production Swift role is set up, otherwise DCops may miss
  some provisioning issues (like in the case of the task).
  IIUC adding the production role directly is not affecting
  the Swift cluster so other manual actions are need to add it
  to the ring.
- Due to an issue with Supermicro and HW-Raid / extra-disks,
  the OS right after reimage doesn't recognize the new disks.
  This in turn causes the puppet no-op run to fail, but (at this
  is my speculation) some state is added to Puppetdb anyway.
- To clear the state, I've done something like T371400#10236705
  multiple times, otherwise subsequent reimages failed.

The idea of this change is to add an extra step in rollback() to
clean up puppetdb state (or at least attempt to do it) if
any previous attempt to populate puppetdb was kicked off.
In theory this should be safe-enough to run, and it should
hopefully avoid DCops to stall/ping us to get the state
properly cleaned up.

Bug: T371400
Change-Id: I2096528788f858253d5812de622b4088fa17e1fa