sre.hosts.provision: wait for 5 mins after rebooting a Supermicro
Sometimes 3 minutes it is not enough for an host to go through
a reboot, and the bios settings applied in the first round
(quiet boot, EFI/BIOS, etc..) are not picked up by
the second round's bios settings refresh (via Redfish).

Increaes to 5 minutes seems to solve the problem. I know that
we need a better solution for this, but it is all I have at
the moment :D

Change-Id: If7e1a55e8d013eedf219910b3edd61ffb2a61a7b