EFI don't remove override twice
As part of the debian-installer, grub-install both installs the grub
boot loader in the EFI partition and creates an EFI boot entry to point
to the installed boot loader. The new EFI boot entry is configured first
in the boot order. However, the call to override the boot target during
the debian-installer, via Redfish, causes the EFI boot changes to be
lost on reboot.

Change-Id: I8759d663fa4a4c2c52f79efb72ed5854bc518a82