sre.hosts.{dhcp,reimage}: force tftp as default option
To simplify the handling of reimages and related operations,
we are setting the usage of TFTP only in DHCP configuration
as default, allowing to use HTTP as opt-in argument that
the user needs to provide.

The TFTP only configurations are needed to workaround some
of the issue highlighted in T363576, happening only on
servers equipped with 10Gbps NICs. Since the usage of the flag
was added while experimenting in production the new feature,
we promote it now to the default to avoid confusion in people
trying to reimage.

Bug: T363576
Change-Id: I334eab30c6ed1e1cefec97a58111cd10f6beeb0b