sre.wdqs.data-transfer: improve graph type checks
We already directly check the graph type being
transferred; I don't think we need the
check_hosts_are_valid check
for the transfer anymore. But if we do in the future
re-adding this will be trivial.

Also fix a bug where a categories transfer would set
the data_loaded flag to "categories". In actuality
we don't want to do this; we only want to set the flag
if the transfer type is main, scholarly, or commons
(i.e. is not `categories`)

Bug: T376150
Change-Id: Icbb66af68b1b513cd16cfed2e24244cb4b9dd857