move-vlan: Don't fail if there is nothing to do
Currently the move-vlan cookbook fails if everything is fine (e.g. the
server does not need to be moved). This is a bit counter intuitive when
doing large numbers of reimages where we know that the host IPs don't
matter (e.g. kubernetes nodes) and we can do move-vlan without issues.

Change-Id: Id08c10f40391479d44295f6db72485c9bbcfbc2f