reimage: change check to find physical interface if IP is on bridge
Similar to I3baca91d678be3baa16b794eaecf01d3468c7703 if a host
primary IP is attached to a bridge device we need to do a further
lookup to find the physical port that belongs to it to work out
what switch the server is connected to.

The logic for this was in the reimage cookbook, but depended on
nbiface.connected_endpoints to be equal to None if that was the
case, after which it does the second query.  Since the Netbox
model change to support N:1 endpoints, we check for
connected_endpoints[0], which if it is none throws an error.

Adjusting the logic to check the interface type before we
try to get the connected_endpoints.

Bug: T383207
Change-Id: Ib18de45213f3c131d0698de5978c1e344790a26c