k8s.pool-depool-node: Add support to downtime/remove downtime
This adds support to downtime nodes prior to depooling and remove a
downtime after pooling. Removing downtime is a bit cumbersome since we
have to store the downtime ID somewhere in order to be able to remove
alertmanager silences (see T395032).
The workaround for this is to print the pool command (including the
downtime ID) and also post it to phab if --task-id is provided.

Taking the chance to refactor creating the downtime duration arguments
into a function provided by cookbooks.sre.hosts.downtime.

Bug: T341984
Change-Id: I61d1ac77e138edf53d61ec6ececf236d35d8bb83