Network: add qos and sflow config for configure-switch-interfaces
This appends additional commands to those that are pushed to network
devices when the sre.network.configure-switch-interfaces cookbook is
run.  The goal is to reduce warnings mails from our automatic
Homer-based diff checker, which currently fire as Homer does other
stuff apart from just enable the interface when it is added in
Netbox.

As mentioned on task the solution probably isn't perfect but it may
help for the time being.

Bug: T379549
Change-Id: Ife3e1b44fd3cbb3caa4e61a02659acaa72dd054c