sre.hardware.upgrade-hardware: Mention possibly long run time
On some servers the upload may take quite some time to upload the
firmware to the BMC (3 yo Dell 440 with old 5.x firmware), which
led me to believe the cookbook might have been stuck.

So add a note to warn that taking up to half an hour is expected
in some cases.

Bug: T385873
Change-Id: I6e4a2abcbcf7e8ae2d5c7f27291dac632660a97f