sre.hosts.provision: add logic to set PXE for Supermicro
After some discussion with DCops, we decided to set PXE to
only one NIC at the time, to avoid booting issues as much
as possible. We do set either "Legacy" or "EFI" for all
devices at the moment, meanwhile it seems that a better config
would be to set "PXE" to the only NIC supposed to PXE boot.

The Legacy/BIOS mode use case is simpler, since the BIOS
settings for the NICs accept "Disabled" and "PXE", meanwhile
for UEFI is different since the values are three: "PXE", "EFI"
and "Disabled". The "Disabled" value is the only constant one,
"PXE" and "EFI" seem to depend from the firmware version.

Bug: T387577
Change-Id: I754491fd6fa077011340662bbe60ab6eb628cc7a