cdn: Unify ats/haproxy/varnish upgrade cookbooks
Our needs for upgrading CDN components only differ when it comes to how
we deploy them (e.g. grace periods) and what services/packages we want
to upgrade. To that end, they should all be pretty much the same
cookbook so this commit unifies most of the logic.

The HAProxy cookbook was hit the hardest but it deviated the most. We
can simplify it by using the SRELBBatchRunnerBase for pooling/depooling,
and Spicerack's apt_get module already handles the
--force-confdef/--force-confold options by default. Therefore, we can
just do away with our own implementation and lean on that.

Why not a single cookbook?

The pseudo-declarative nature of cookbooks make it really awkward to set
the conditionals that we'd need for each of the services. For instance,
if we do anything to Varnish that needs a restart, we'd need to
instantly increase the grace period. However, we can't easily set that
since it has to be in the SREBatchBase-based class and we'd only be able
to detect whether it needs to be increased/decreased "downstream" in the
SRELBBatchRunnerBase class.

In the end, the complications arising from all the plumbing we'd need to
accomplish for unifying this into one cookbook isn't really worth the
savings.

Change-Id: I771bcfe1d22d16526baa249df8d65a3dd8f7fe41