sre.hosts.provision: try Supermicro BMC passwords automatically
A source of confusion for the Supermicro BMC setting is most of
the time what password is needed to allow the cookbook to add
the WMF specific settings to a brand new host.

We have currently some use cases:
1) The host is brand new, it runs the custom WMF firmware that
   sets the password to calvin (as it happens for Dells).
2) The host is brand new, but without the custom firmware.
   The BMC password is the one stated on the server's label.
3) The host is already provisioned, and the password is
   the WMF mgmt one.

It is feasible then to automate the test of all the above cases,
asking for the BMC label password only when needed (as opposed
to always). This should allow the operator of the cookbook
to be less confused on what password to use.

Bug: T386946
Change-Id: Id6ea0a29bb9a5d6f3c6f35f5fa2d84e2cc99b791