sre.hosts.provision: add a warning for ipv6 disabled
For some obscure reasons that I still haven't got, sometimes
we get a Supermicro server with its BMC IPv6 settings disabled.
This in turn causes our IPv6 StatelessAddressAutoConfig to raise
an exception when issuing settings to Redfish, with a cryptic
message.

I tried to find a way to explicitly check if the BMC's IPv6
is enabled or not, but I haven't found anything good.

Add a more meaningful error message, with the only known
workaround, to help whoever runs provisioning.

Bug: T389950
Change-Id: Ie41e04f010ecba84ce51ab55dbe516f02108323b