sre.hosts.move-vlan: improve grep reports when renaming
If a host is being renamed, it may happen that its DNS record
is gone when move-vlan runs. Part of its checks are related
to grepping occurrences of the host's harcoded IP address
in various repos (not fqdn). If there are no A/AAAA records for
the host, the grep will be too broad and a ton of files will
be wrongly showed up to the user.

Skip check_patterns_in_repo if self.patterns is an empty list.

Bug: T392729
Change-Id: I7ca372a5542e2c30ef0f7dac879b840aba145d28