swift: add a check-dbs cookbook to check swift container dbs
We've sometimes seen failures where swift discovers all its copies of
a container DB are corrupt at once, leading to outages. This cookbook
allows us to check the sqlite3 databases of an individual container to
find any that need repairing.

Backend and frontend selection functions split out into __init__.py as
they're likely to be useful again - there are more cookbooks to do
with handling the container DBs (e.g. to VACUUM them), for which these
will be resuable.

This has been tested with test-cookbook, including:
 * specification of frontend and backend hosts
 * error behaviour when container doesn't exist
 * restricting checking to only one DC

test-cookbook has also been used to check that the changes to
remove-ghost-objects haven't broken it

Change-Id: I5bef77e7b7e3d6a8d12c5130ee03671ca309932a