cookbooks/sre: use A:dnsbox everywhere instead of -rec/-auth
There is really no dns-rec or dns-auth anymore, as distinct boxes. There
is just one DNS role (dnsbox) that does all of the things. So while the
current references to dns-rec and dns-auth will still work, they may
give the false impression that these are distinct things.

I noticed this while working on another cookbook so this CR fixes that,
so simply refer to A:dnsbox everywhere. Note that this is just for the
cumin aliases -- rec and auth are still distinct services.

Change-Id: I4eb0cf805bf18ba21c71643b59ca8d9f3926ec41