sre.wipe-cluster: Ask user to confirm target k8s version
We had an issue where the puppet run on the apiservers would not fail
although it could not install all packages properly. That led to the new
cluster being initialized with an old apiserver version.

Add a check (e.g. ask the user) if the version the cluster was
initialized with is the expected target k8s version.

Fix some pylance warnings along the way.

Bug: T397148
Change-Id: Ia67a30bdd88cad7f8fb75b7fdb8e55dd7134c7ae