reimage: add dhcp MAC address support for physical hosts
Prior to this change we used information provided by dhcp option 82 to
correlate a dhcp request with the host being re-imaged. After this
change we add a new flag --use-mac which tries to use the mac address of
the PXE boot interface as the key to correlate the dhcp request.

The mac address is obtained from netbox, if a manual entry exists. If
not, redfish is queried, and finally puppetdb is used if a mac is not
found in redfish.

Change-Id: Ia22987191e6e3b43fe0efed9d99f1c22583ad551