sre.hosts.provision: add custom settings for Supermicro
We are onboarding new Supermicro hosts:
- ML nodes, UEFI only with a different way of PXE booting.
  Since they are bigger nodes, the firmware allows to use
  the FixedBootOrder API, and it is not required to set
  PXE to the NIC port with LinkUp since the UEFI PXE boot
  will test all the ones available.
  After various tests I decided to just make sure that "Hdd"
  is the first device listed in the FixedBootOrder API.
- Single socket/cpu sretest host, that probably have a different
  firmware and some BIOS property have been renamed.
  On the BMC Network configuration side, the FQDN option is now
  not writable anymore, but from my tests it doesn't seem to
  affect any functionality.

Part of the changes are related to allow the first batch of nodes
to be configured correctly, skipping the configuration of the PXE
NIC completely (but allowing to properly set the UEFI Http Boot
configurations).

I haven't found a better way to apply settings other than using
the model type, that is complicating a little the code because
more if/else blocks are needed.

Bug: T394357
Change-Id: Ie9aa22a32c4fbcaece71243115ea274b02fac843