sre.hosts.provision: update cookbook for Dell iDRAC 10
In newer Dell hosts the SCP config that we dump/modify/upload
doesn't contain anymore a section/component related to the NIC's
configuration, but it seems relying only on BIOS options.
We already started to do the same process when setting UEFI PXE,
but the code meant to find the list of NICs still relies on
the old format of the SCP dump.

Add a new idrac 10 config since a lot of options have been renamed,
and some of them may or may not be available (so far we have
only tested single CPU hosts though).

Add also support to configure a RAID device, instead of an HDD one,
since new BOSS-based hosts will need it.

Bug: T392851
Change-Id: I26e22033afa3bb4a9dce459aed0dbe9fe132ea25