sre.gitlab.failover: manage home_page_url in cookbook
This change adds a call to gitlab-rails console to update the home_page_url
for the switch_from_host. This makes sure the old source host
uses the proper home_page_url. This step was handled by gitlab-settings
previously, so adding it to the cookbook reduces the number of
manual steps.

The home_page_url for the new switch_to_host is handled by the backup-
restore script already.

Bug: T400695
Change-Id: Idf76bb39de9b44b6ab9a2d68e0be351aae4f3e06