sre.hosts.provision: fix check for idrac10
In my last patch I added a check for the Redfish generation value
on the Dell class, to establish if the target host runs idrac 10+.
This check assumes that Redfish can be queried and used, but
it cannot happen until DHCP and other settings have been applied.

Set the check in the _config_host() function.

Bug: T392851
Change-Id: I599edb44f3d1516338644ff6f9853d7c261d802f