sre.hosts.provision: expand Supermicro models with no PXE devs in BIOS
We are trying to fix some inconsistency in BIOS settings for ml-serve
hosts with GPUs, that were racked a year ago and I don't see traces
in Phabricator of special patches or fixes. The BIOS doesn't show
any key/setting with "LAN" as substring, indicator that the setting
is related to a NIC (we then use it to set PXE on it).

In recent ml hosts with GPUs, it came up that Supermicro implemented
a single generic BIOS setting to PXE boot from all devices when needed.
The major difference is that the recent GPU hosts are UEFI only,
meanwhile the older ones can also work in legacy mode. The CSMSupport
(to allow MBR-based systems in UEFI mode) seems not to be tunable,
so make it optional in the code as well.

Update the provision cookbook with the new model, so that we can
experiment with provisioning.

Bug: T401964
Change-Id: Iee74437451c8112161f0180671cb1638890fa8e7