sre.host.provision: move WebServer.1#HostHeaderCheck as optional
On ml-serve1008 it seems that WebServer.1#HostHeaderCheck is not
available (checked via spicerack-shell). It may be related to
its iDRAC firmware, but it seems ok to just make the field optional
when creating the Dell's config changes.

A more precise and long-term fix could be to work on Spicerack to
avoid new settings to be added to the config if not present, but
we can use the aforementioned interim solution for the time being.

Bug: T401964
Change-Id: If4f413b9bb4ba8dcc2adff637df485910f0a4faa