sre.hsots.provision: add sys-121c-tn2r-configg as special case
The dse-k8s-worker1014 host has the BIOS option for ConsoleRedirection
named COM1ConsoleRedirection, so it needs to be put in
the appropriate list.

Moreover, it seems that it doesn't list the ADMIN user. We usually set
its password with the root one, but if not present we can simply
skip it. Refactor the logic for ADMIN/root password a bit to take
the above into account.

Bug: T399779
Change-Id: I9c12caa7c2e9521aad557c13f38dc0f787c3b039