provision: on reboot wait for bios attrs
Prior to this change we considered the boot complete when the boot state
changed to a state signaling the box is booting the OS. But, we also
want to ensure the BIOS attributes are available to read from Redfish.
After this change we poll for the BIOS attributes, before considering
the reboot complete.

Change-Id: Id4989421630ea5e7ff88b54cca15b0521b6aeac5