sre.hardware.upgrade-firmware: fix ssd upgrade
In the IDRAC and BIOS use cases, if the target version retrieved
by _update is equal to the current version of the firmware, we return
"True" as sign of success. In the ssd use cases we don't,
leading to unnecessary cookbook failures.

Instead of an early return, leave the final check to the code that
follows, since it inspects every drive checking its firmware
version.

Bug: T392851
Change-Id: I8b0f63c8ec3094e5487d64cab53eb4a245c71ab3