sre.hosts.provision: remove boot order config in UEFI for Dells
In https://gerrit.wikimedia.org/r/c/operations/cookbooks/+/1173335
I introduced a bug, that is triggered when a Legacy host is moved
to UEFI. The Legacy settings don't include the UefiBootSeq entry
in the BIOS, so the fetch fails.

In T406964 the first Dell config J node was flipped to UEFI,
but despite a lot ot tests/upgrades we were not able to get
a disk option listed in the boot option (before d-i).
Since the Debian installer takes care of setting the right UEFI
boot order, there is no point in trying to configure one
during provisioning. This removes a lot of headaches and
compatibility issues that we have been trying to solve
recently.

Bug: T406964
Change-Id: Id166ff2224f27e6e85fab85278d7fddc4651dcfe