wdqs.data-transfer: --force behavior now mandatory
When --force is set it wipes out relevant files on the
destination host, such as the journal file.

Without the flag, if the file is already present on the dest,
the transfer will not overwrite it (basically leading to a silent failure to transfer).

Since we basically always want the --force behavior,
let's make it default and remove the option entirely.

Bug: T408163
Change-Id: I7b626a033637d9ab8a070c009f332ab3e5eef613