Add the sre.hosts.powercycle cookbook
Add the powercycle cookbook as alternative to the reboot-single,
to offer an alternative for certain use cases where we need
to interact with the BMC.

Refactor the _reboot_chassis() method in the provision cookbook
to a more generic one, usable by both cookbooks.

While working on ml-serve2001 to test the cookbook, I noticed
that the 'BootProgress' attribute is not available, so I added
a break in the code for that specific use case. We'll fall
back to wait and test the presence of the BIOS attributes, so
it should be ok as well (less complete and defensive of course).

Another weird thing that I noticed is that ml-serve2001 doesn't
accept 'GracefulRestart' as ResetType, but 'Powercycle'. I have
seen this issue before, worth to follow up in a task, but for
this use case we can just use ForceReset.

Change-Id: I35ba3979f41fef7c81079635968d31619c81bbf3