reimage: force --no82 if device is connected to Nokia switch
Nokia support for inserting the L2 port name when it adds Option 82
information to a DHCP DISCOVER packet did not make it into release
v25.10.1 as expected.  Further we are not using the Nokia switches
as L3 gateway for connected hosts as of yet (see T409286) which
means we cannot use the relay function on them (and they do not
support inserting Option 82 info outside of when relaying).

This change forces the --no82 option to on if the host being reimaged
is connected to a Nokia switch.

Bug: T410751
Change-Id: I51fca4fc9d1204a59538d40afbfdeb5a67851546