hadoop.reboot-workers: make host override smarter
[CONTEXT]

There's basically two ways we use this cookbook:
(1) running it on a whole cluster
(2) running it on a subset of a cluster with a host override

(2) is generally done if we got partially through
a reboot with (1) and then hit an error, or because
there's some reason we only need to reboot a few hosts
in the fleet.

The override logic before this patch was very simple,
it would just reboot all the provided hosts per the
batch size and not care about ensuring that journal
nodes were only restarted one host at a time.

This required the operator to manually curate the
override cumin query to ensure it only contains a single
journal node. Besides being a massive foot-gun, this also
meant that it could take multiple invocations to get through
a whole fleet since we could only have one journal node
present at a time in each cumin query.

[SOLUTION]

Steal the existing logic for the non-override code path, and just
do the same thing in the override codepath.

We've also recently switched to get_subset instead of intersection,
so continue the usage of that new pattern here.

Also as a freebie add a --dry-run mode that won't mutate the cluster,
but instead will just log.

Bug: T411568
Change-Id: If6669440227ab4f32c2a1151b9279f3b3e1570c9